+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `PackageMetadata::iter_changelog_entries()` for iterating over changelog entries without collecting them into a list.
+
 ## 0.28.0
 
 ### Fixed